Fix bug #577 [^]] not working C.UTF8
In regconp()-->bra() i.e bracket handling, in the

	if (complicated && mbcoll())
	{

ie (i.e UTF8) when handling ']' in a negative pattern class, i.e [^]...]
the ']' is mis handled and then enter wrong elements in the collate table.
It mislead the regnexec()... not detecting the ']'.

The original code

else if ((c == ']')
{ ....

The fix

else if ((c == ']') && !(neg && (first == pp)))
{ ....

If we are in the neg case [^ and if first is '[' and pp is at the first
position don't do this. I then fallback in the normal processing effectivly
enter  the ']' on the negate class.